requirements.txt
fire
tqdm
nltk
fasttext
py-rouge
pandas
unidecode
torch==1.10.0
transformers==4.18.0
scikit_learn==1.0.2
pytorch-lightning==1.6.2
bert-score==0.3.11